都市オブジェクトの構造体を "Objectify" する仕組み (#42)
Rustの構造体として読み込まれたCityGMLの都市オプジェクトを、より汎化されたデータ型での表現（JSONのようなものをイメージしてもらえれば）に変換できるようにする。これによって後段での加工や、出力ドライバでの扱いができるようになる。

今後も色々とブラッシュアップしていくと思いますが、大枠はこれでいいのでいったんPR投げます。

<img width="484" alt="object"
src="https://github.com/MIERUNE/nusamai/assets/5351911/49fa57a7-1b6c-4c89-97ff-9b5904c23045">

Closes #40

+++ nusamai-plateau/src/models/relief.rs
@@ -10,7 +10,7 @@ pub struct ReliefFeature {
     name: Option<String>,
 
     #[citygml(path = b"dem:lod")]
-    lod: Option<i32>,
+    lod: Option<u64>,
 
     #[citygml(path = b"dem:reliefComponent")]
     relief_component: Vec<ReliefComponent>,
@@ -44,7 +44,7 @@ pub struct TINRelief {
     #[citygml(path = b"gml:name")]
     name: Option<String>,
 
-    #[citygml(auto_geom = b"dem")]
+    #[citygml(geom = b"dem")]
     geometries: GeometryRef,
 }
